+++ entity/src/main/resources/db/migration/tables/tenant/V1.9__update_t_tenant_project.sql
@@ -1,0 +1,12 @@
+-- ------------------------------------------------
+-- Version: V1.9
+-- Description: Update t_Tenant_Project table
+-- ------------------------------------------------
+ALTER TABLE tenant_project
+DROP CONSTRAINT fk_tenant_project_project;
+
+ALTER TABLE tenant_project
+ADD CONSTRAINT fk_tenant_project_project
+    FOREIGN KEY (project_id)
+    REFERENCES t_Project(id)
+    ON DELETE RESTRICT;